Enabled PSO to take arguments for cost function
The PSO function can now take arguments and it will pass these then to
the cost function.
NOTE: The cost function will need to have *args in the function
definition. The arguments will be in the form of a tuple so will need to
be unpacked.

Set the default velocity limit scale in the params section.