Implemented a version of binary PSO
This version of binary PSO is from the paper Kennedy, James, and Russell C. Eberhart. "A discrete binary version of the particle swarm algorithm." 1997 IEEE International conference on systems, man, and cybernetics. Computational cybernetics and simulation. Vol. 5. IEEE, 1997.

It uses a sigmoid function and a uniform random variable to decide if the bit will be flipped. For more information see the paper referenced above.

Binary PSO is enabled by setting the parameter BPSO to 1 (default = 0).